[feat]:CategoryCell Layout 파일 분리 및 MARK 주석 작성
